[♻️refactor]: Home 페이지의 MainContent 파일 분리
- components/Pages/Home 디렉토리 내의 파일로 분리함
